# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Chores API","description":"The the hassle out of chore a\u00a0assignments.","termsOfService":"https:\/\/chores.cosmigrams.com\/terms\/","contact":{"email":"sutabi@gmail.com"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"servers":[{"url":"http:\/\/localhost:8000\/api\/v1","description":"Chores Api Server"}],"paths":{"\/users":{"get":{"tags":["user"],"summary":"Get list of users under your account","responses":{"200":{"description":"object of your account details","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}}},"security":[{"passport":[]}]},"put":{"tags":["user"],"summary":"Update a user","parameters":[{"name":"id","in":"path","description":"ID of user to update","required":true,"schema":{"type":"string"}}],"requestBody":{"request":"User","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"responses":{"200":{"description":"Successful in updating user"},"400":{"description":"Validation error"},"404":{"description":"User not found"}},"security":[{"passport":[]}]},"post":{"tags":["user"],"summary":"Create a user","requestBody":{"request":"User","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"responses":{"200":{"description":"Successful in updating user"},"400":{"description":"Validation error"},"404":{"description":"User not found"}}}},"\/users\/me":{"get":{"tags":["user"],"summary":"Get your account details","responses":{"200":{"description":"object of your account details","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}}},"security":[{"passport":[]}]}},"\/users\/{id}":{"get":{"tags":["user"],"summary":"Get a user under your account","parameters":[{"name":"id","in":"path","description":"ID of user to fetch","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"object of your account details","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"400":{"description":"Missing ID"},"404":{"description":"Record not found"}},"security":[{"passport":[]}]}},"\/workers":{"get":{"tags":["worker"],"summary":"Get list of workers under your account","responses":{"200":{"description":"array of your workers","content":{"application\/json":{"schema":{"type":"array","items":{"$ref":"#\/components\/schemas\/Worker"}}}}}},"security":[{"passport":[]}]},"post":{"tags":["worker"],"summary":"Create a worker","requestBody":{"request":"User","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/WorkerCreate"}}}},"responses":{"200":{"description":"Successful in updating worker","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Worker"}}}},"400":{"description":"Validation error"}},"security":[{"passport":[]}]}},"\/workers\/{id}":{"get":{"tags":["worker"],"summary":"Get a worker under your account","parameters":[{"name":"id","in":"path","description":"ID of worker to fetch","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"object of your account details","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Worker"}}}},"400":{"description":"Missing ID"},"404":{"description":"Record not found"}},"security":[{"passport":[]}]},"put":{"tags":["worker"],"summary":"Update a worker","parameters":[{"name":"id","in":"path","description":"ID of worker to fetch","required":true,"schema":{"type":"string"}}],"requestBody":{"request":"User","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/WorkerUpdate"}}}},"responses":{"200":{"description":"Successful in updating worker","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Worker"}}}},"400":{"description":"Validation error"}},"security":[{"passport":[]}]}}},"components":{"schemas":{"WorkerCreate":{"title":"Worker Create Model","description":"The worker \/ user model to create a new worker","required":["name","email","password"],"properties":{"name":{"title":"Full Name","description":"Entire name, first, last, middle","type":"string","maxLength":255,"example":"John Doe"},"email":{"title":"Email","description":"Email","type":"string","format":"email","maxLength":255},"password":{"title":"Password","description":"Password at least 6 characters","type":"string","format":"password","maxLength":255,"minLength":6},"password_confirmation":{"title":"Password","description":"Same as password field","type":"string","format":"password","maxLength":255,"minLength":6},"phone":{"title":"Phone","description":"Phone number","type":"string","maxLength":60,"example":"(619) 553-2123"},"pushover_key":{"title":"Pushover Key","description":"Integration key for Pushover Push Notifications","type":"string","maxLength":255,"example":"vb34tghd34ger34g"},"notice_by":{"title":"Notification Setting","description":"How does the user want to receive notifications?\nPossible Values:\n- 0 - No Notifications\n- 1 - Email Notifications\n- 2 - Sms Notifications\n- 3 - Push Notifications\n","type":"integer","maxLength":16,"example":0},"sms_by":{"title":"Text Message (SMS) Setting","description":"If the user's notification is set to 'sms', they have the option to receive via\nSMS gateway, or email to sms ?\nPossible Values:\n- 0 - No Notifications\n- 1 - Text To SMS - This just uses your cellphone number and sends a true SMS\n- 2 - Email To SMS - This uses your cellphone number and appends it to a SMS Gateway i.e {n}@tomail.net\n","type":"enum","maxLength":16,"example":0},"sms_email":{"title":"Text Message (SMS) Email","description":"If the user has selected to be notified by SMS and using their SMS email, then provide the address here","type":"email","maxLength":255,"example":"{n}@tmomail.net"}},"type":"object"},"WorkerUpdate":{"title":"Worker Update Model","description":"The worker \/ user model to update","properties":{"name":{"title":"Full Name","description":"Entire name, first, last, middle","type":"string","maxLength":255,"example":"John Doe"},"email":{"title":"Email","description":"Email","type":"string","format":"email","maxLength":255},"password":{"title":"Password","description":"Password at least 6 characters","type":"string","format":"password","maxLength":255,"minLength":6},"phone":{"title":"Phone","description":"Phone number","type":"string","maxLength":60,"example":"(619) 553-2123"},"pushover_key":{"title":"Pushover Key","description":"Integration key for Pushover Push Notifications","type":"string","maxLength":255,"example":"vb34tghd34ger34g"},"notice_by":{"title":"Notification Setting","description":"How does the user want to receive notifications?\nPossible Values:\n- 0 - No Notifications\n- 1 - Email Notifications\n- 2 - Sms Notifications\n- 3 - Push Notifications\n","type":"integer","maxLength":16,"example":0},"sms_by":{"title":"Text Message (SMS) Setting","description":"If the user's notification is set to 'sms', they have the option to receive via\nSMS gateway, or email to sms ?\nPossible Values:\n- 0 - No Notifications\n- 1 - Text To SMS - This just uses your cellphone number and sends a true SMS\n- 2 - Email To SMS - This uses your cellphone number and appends it to a SMS Gateway i.e {n}@tomail.net\n","type":"enum","maxLength":16,"example":0},"sms_email":{"title":"Text Message (SMS) Email","description":"If the user has selected to be notified by SMS and using their SMS email, then provide the address here","type":"email","maxLength":255,"example":"{n}@tmomail.net"}},"type":"object"},"User":{"title":"User model","description":"User model","required":["name","email","password"],"properties":{"id":{"title":"User ID","description":"UUID of user","type":"string","maxLength":255,"example":"f80cf05a-9ef3-4781-9a05-728079e360ae"},"name":{"title":"Full Name","description":"Entire name, first, last, middle","type":"string","maxLength":255,"example":"John Doe"},"email":{"title":"Email","description":"Email","type":"string","format":"email","maxLength":255},"phone":{"title":"Phone","description":"Phone number","type":"string","maxLength":60,"example":"(619) 553-2123"},"pushover_key":{"title":"Pushover Key","description":"Integration key for Pushover Push Notifications","type":"string","maxLength":255,"example":"vb34tghd34ger34g"},"notice_by":{"title":"Notification Setting","description":"How does the user want to receive notifications?\nPossible Values:\n- 0 - No Notifications\n- 1 - Email Notifications\n- 2 - Sms Notifications\n- 3 - Push Notifications\n","type":"integer","maxLength":16,"example":0},"sms_by":{"title":"Text Message (SMS) Setting","description":"If the user's notification is set to 'sms', they have the option to receive via\nSMS gateway, or email to sms ?\nPossible Values:\n- 0 - No Notifications\n- 1 - Text To SMS - This just uses your cellphone number and sends a true SMS\n- 2 - Email To SMS - This uses your cellphone number and appends it to a SMS Gateway i.e {n}@tomail.net\n","type":"integer","maxLength":16,"example":0},"sms_email":{"title":"Text Message (SMS) Email","description":"If the user has selected to be notified by SMS and using their SMS email, then provide the address here","type":"email","maxLength":255,"example":"{n}@tmomail.net"},"created_at":{"title":"Date Created","description":"When the user was created","type":"datetime","maxLength":16,"example":"01\/01\/2022 05:00:00"},"updated_at":{"title":"Date Updated","description":"When the user was last updated","type":"datetime","maxLength":16,"example":"01\/01\/2022 05:00:00"}},"type":"object"},"Worker":{"title":"User model","description":"User model","properties":{"id":{"title":"Worker ID","description":"UUID of Worker","type":"string","maxLength":255,"example":"f80cf05a-9ef3-4781-9a05-728079e360ae"},"owner_id":{"title":"Worker Owner","description":"UUID of worker's owner id","type":"string","maxLength":255,"example":"f80cf05a-9ef3-4781-9a05-728079e360ae"},"user_id":{"title":"Worker User","description":"UUID of worker user id","type":"string","maxLength":255,"example":"f80cf05a-9ef3-4781-9a05-728079e360ae"},"name":{"title":"Full Name","description":"Entire name, first, last, middle","type":"string","maxLength":255,"example":"John Doe"},"user":{"title":"User","description":"Worker's user data","type":"object","$ref":"#\/components\/schemas\/User"},"created_at":{"title":"Date Created","description":"When the user was created","type":"datetime","maxLength":16,"example":"01\/01\/2022 05:00:00"},"updated_at":{"title":"Date Updated","description":"When the user was last updated","type":"datetime","maxLength":16,"example":"01\/01\/2022 05:00:00"}},"type":"object"}},"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","securityScheme":"passport","flows":{"password":{"authorizationUrl":"http:\/\/localhost:8000\/oauth\/authorize","tokenUrl":"http:\/\/localhost:8000\/oauth\/token","refreshUrl":"http:\/\/localhost:8000\/oauth\/refresh","scopes":[]}}}}},"tags":[{"name":"user","description":"Operations about the user"},{"name":"worker","description":"Operations of the worker"}],"externalDocs":{"description":"Find out more about Chores","url":"https:\/\/chores.cosmicgrams.com\/"}}